converted floats to doubles and fixed the +- button when a decimal is at the end

+++ Calculator/CalculatorLib/CalculatorLib.cpp
@@ -107,22 +107,22 @@ std::string infixToPostfix(std::string &a_sInfix) //Converts an infix notation s
 	return sPostfix;//Returns the final postfix notation of the equation.
 }
 
-double rpnEvaluate(const std::string &a_c_sPostfix)//A function that takes in a string as an argument that should be a postfix equation and returns a double that is the answer.
+long double rpnEvaluate(const std::string &a_c_sPostfix)//A function that takes in a string as an argument that should be a postfix equation and returns a double that is the answer.
 {
 	//Current operations: +, -, *, /, ^.
 	std::istringstream iss(a_c_sPostfix);//Creates a buffer from the a_c_sPostfix string.
-	std::stack<double> sdStack;//Creates a stack that ends up containing one element which is the answer.
+	std::stack<long double> sdStack;//Creates a stack that ends up containing one element which is the answer.
 	std::string sToken;//Creates a string that is used to temporarily hold the value of each set of characters from the a_c_sPostfix string.
 	while (iss >> sToken)//Iterates through each part of the a_c_sPostfix string making each set of characters equal to sToken until it reaches the end.
 	{
-		double dToken;//A temporary value to check whether the token can be passed as a double or not.
+		long double dToken;//A temporary value to check whether the token can be passed as a double or not.
 		if (std::istringstream(sToken) >> dToken)//If the token can be passed as a double return true.
 			sdStack.push(dToken);//Pushes the token which is a double to the top of the stack.
 		else//If the token can't be passed as double.
 		{
-			double dNumberTwo = sdStack.top();//Creates a double variable that is equal to the top of the stack which defaults to 0.
+			long double dNumberTwo = sdStack.top();//Creates a double variable that is equal to the top of the stack which defaults to 0.
 			sdStack.pop();//Removes the top of the stack. 
-			double dNumberOne = sdStack.top();//Creates a double variable that is equal to the top of the stack which defaults to 0.
+			long double dNumberOne = sdStack.top();//Creates a double variable that is equal to the top of the stack which defaults to 0.
 			sdStack.pop();//Removes the top of the stack.
 			if (sToken == "*")//Checks to see if the token is exactly *.
 				sdStack.push(dNumberOne * dNumberTwo);//Pushes the answer of dNumberTwo * dNumberTwo onto the sdStack.
