+++ Calculator/CalculatorLib/CalculatorLib.h
@@ -19,7 +19,7 @@ bool isNumber(const std::string &a_c_sToken);//Returns true if the last characte
 bool isOperation(const std::string &a_c_sToken);//Returns true if the passed in token matches one of the operations.
 int operationRank(const std::string &a_c_sToken);//Returns an int equal to the rank the passed in token operator has over other operators in accordance to BODMAS.
 std::string infixToPostfix(std::string &a_c_sInfix);//Converts an infix notation string and returns a postfix notation version.
-double rpnEvaluate(const std::string &a_c_sPostfix);//A function that takes in a string as an argument that should be a postfix equation and returns a double that is the answer.
+long double rpnEvaluate(const std::string &a_c_sPostfix);//A function that takes in a string as an argument that should be a postfix equation and returns a double that is the answer.
 std::string calculate(std::string &a_c_sInfix);//Uses all the above functions to convert the infix notation to postfix and then evaluate it.
 
 #endif
